unify install from CD logic
