+++ OpenRA.Mods.TS/OpenRA.Mods.TS.csproj
@@ -46,7 +46,6 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Widgets\Logic\TSInstallFromCDLogic.cs" />
     <Compile Include="ShroudPalette.cs" />
     <Compile Include="Activities\VoxelHarvesterDockSequence.cs" />
     <Compile Include="TiberianSunRefinery.cs" />
