+++ mods/cnc/mod.yaml
@@ -137,6 +137,10 @@ ContentInstaller:
 	FilesToCopy: CONQUER.MIX, DESERT.MIX, SCORES.MIX, SOUNDS.MIX, TEMPERAT.MIX, WINTER.MIX
 	FilesToExtract: speech.mix, tempicnh.mix, transit.mix
 	PackageMirrorList: http://openra.net/packages/cnc-mirrors.txt
+	DiskTestFiles: CONQUER.MIX, DESERT.MIX INSTALL/SETUP.Z
+	PackageToExtractFromCD: INSTALL/SETUP.Z
+	ExtractFilesFromCD: speech.mix, tempicnh.mix, transit.mix
+	CopyFilesFromCD: CONQUER.MIX, DESERT.MIX, SCORES.MIX, SOUNDS.MIX, TEMPERAT.MIX, WINTER.MIX
 
 ServerTraits:
 	LobbyCommands
