+++ mods/ra/mod.yaml
@@ -135,6 +135,10 @@ ContentInstaller:
 	InstallerMenuWidget: INSTALL_PANEL
 	TestFiles: allies.mix, conquer.mix, general.mix, interior.mix, redalert.mix, russian.mix, scores.mix, snow.mix, sounds.mix, temperat.mix
 	PackageMirrorList: http://openra.net/packages/ra-mirrors.txt
+	DiskTestFiles: MAIN.MIX, INSTALL/REDALERT.MIX
+	PackageToExtractFromCD: MAIN.MIX
+	ExtractFilesFromCD: conquer.mix, russian.mix, allies.mix, sounds.mix, scores.mix, snow.mix, interior.mix, temperat.mix
+	CopyFilesFromCD: INSTALL/REDALERT.MIX
 
 ServerTraits:
 	LobbyCommands
