+++ mods/ts/chrome/gameinit.yaml
@@ -94,7 +94,7 @@ Background@INSTALL_DOWNLOAD_PANEL:
 			Key: escape
 
 Background@INSTALL_FROMCD_PANEL:
-	Logic: TSInstallFromCDLogic
+	Logic: InstallFromCDLogic
 	X: (WINDOW_RIGHT - WIDTH)/2
 	Y: (WINDOW_BOTTOM - HEIGHT)/2
 	Width: 500
